Code review: move notification consumer into its own package
