+++ ee/control/consumers/notificationconsumer/notify_consumer.go
@@ -1,4 +1,4 @@
-package control
+package notificationconsumer
 
 import (
 	"crypto/sha256"
