+++ ee/control/consumers/notificationconsumer/notify_consumer_test.go
@@ -1,4 +1,4 @@
-package control
+package notificationconsumer
 
 import (
 	"bytes"
